+++ src/components/detail/RentButton.jsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 import styled, { css } from "styled-components";
 
 import RENT_STATES from "@/constant/RENT_STATES";
